+++ CHANGELOG.md
@@ -4,8 +4,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Ideas/Upcoming]
-* 
-* TimeBox
 
 ## [Unreleased]
 ### Added
@@ -13,3 +11,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * BrowseTextBox, a EnhancedTextBox which hosts a browse button with a command
 * SearchTextBox, a EnhancedTextBox which has a search and cancel search button with their commands
 * NumberBox, a text box with additional behaviors which accepts a given number type
+* TimeBox, a text box which provides ways to input a time (TimeSpan) in short or long format